agora s

+++ readDatas.py
@@ -31,8 +31,7 @@ class ReadDatas():
                               break
                         cont+=1
                         print(arq)
-                        loaded_data = np.load('./datas3/' + arq)
-                      
+                        loaded_data = np.load('./datas3/'+arq)
                         shape = loaded_data.shape
                         #print(shape,len(dados))
                         aux = [ loaded_data]
@@ -45,7 +44,7 @@ class ReadDatas():
                         dados.append(loaded_data2)
         total_size = len(dados)  # Suponha que temos 100 amostras
         for i in range(total_size):
-            dados[i] = torch.tensor(dados[i]).permute(1, 0, 2, 3)
+            dados[i] = torch.tensor(dados[i],dtype=torch.float).permute(1, 0, 2, 3)
         return dados
 
 if __name__ == "__main__":
